refactor: CalendarScreen 네비게이션 정리 및 캘린더 기능 개선
CalendarScreen 네비게이션 동작을 전용 메서드로 분리해 가독성을 높였습니다.
캘린더 레이아웃·패딩용 상수를 추가해 UI 일관성을 개선했습니다.
년·월 선택을 위한 피커 오버레이를 구현했습니다.
CalendarGrid가 새 패딩 상수를 사용하도록 수정해 레이아웃을 정리했습니다.
CalendarWidget의 구식 년·월 피커 코드를 제거해 구조를 단순화했습니다.
